requirements.txt
amqp                          ==2.5.2       
anyjson                       ==0.3.3       
asgiref                       ==3.2.3       
beautifulsoup4                ==4.8.2       
billiard                      ==3.6.1.0     
celery                        ==3.1.26.post2
certifi                       ==2019.11.28  
cffi                          ==1.13.2      
chardet                       ==3.0.4       
cryptography                  ==2.8         
Django                        ==3.0.2       
django-bootstrap4             ==0.0.8       
django-celery                 ==3.3.1       
django-ckeditor               ==5.7.1       
django-cors-headers           ==3.2.0       
django-crispy-forms           ==1.8.1       
django-filter                 ==2.2.0       
django-google-analytics-app   ==4.4.0       
django-inline-svg             ==0.1.1       
django-js-asset               ==1.2.2       
django-taggit                 ==1.1.0       
django-tagulous               ==0.14.0      
django-tailwind               ==0.5.6       
django-widget-tweaks          ==1.4.5       
djangorestframework           ==3.9.4       
djangorestframework-simplejwt ==4.3.0       
docopt                        ==0.6.2       
idna                          ==2.8         
importlib-metadata            ==1.3.0       
kombu                         ==4.6.7       
mailchimp                     ==2.0.9       
more-itertools                ==8.0.2       
Pillow                        ==6.0.0       
pip                           ==19.3.1      
psycopg2-binary               ==2.8.2       
pycparser                     ==2.19        
PyJWT                         ==1.7.1       
pyOpenSSL                     ==19.1.0      
python-dateutil               ==2.8.0       
pytz                          ==2019.1      
requests                      ==2.22.0      
setuptools                    ==44.0.0      
six                           ==1.12.0      
soupsieve                     ==1.9.5       
sqlparse                      ==0.3.0       
structlog                     ==19.2.0      
urllib3                       ==1.25.7      
vine                          ==1.3.0       
zipp                          ==0.6.0